ml/requirements.txt
# Core ML Libraries
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
pytorch-geometric==2.4.0
torch-scatter==2.1.2
torch-sparse==0.6.18

# Data Processing
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.3.2
scipy==1.11.4

# Deep Learning
transformers==4.36.2
huggingface-hub==0.20.2

# Medical Imaging
nibabel==5.2.0
SimpleITK==2.3.1
pydicom==2.4.4
monai==1.3.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1

# Graph Processing
networkx==3.2.1
igraph==0.11.3

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# Jupyter
jupyter==1.0.0
ipykernel==6.29.0
ipywidgets==8.1.1

# Model Export
onnx==1.15.0
onnxruntime==1.16.3

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0
